+++ client/components/index.ts
@@ -1,0 +1,3 @@
+export * from './forms';
+export * from './layout';
+export { default as GoogleIcon } from './GoogleIcon';
