+++ client/components/layout/index.ts
@@ -1,0 +1,2 @@
+export * from './HomeScreen';
+export * from './SettingsScreen';
